myapp/requirements.txt
django-cors-headers
langchain
python-dotenv
openai
pypdf
sentence_transformers
faiss-cpu
#faiss-gpu
python-dotenv
python-telegram-bot
tqdm